Move CommitRef and TaskRef to models, use toRef().
