+++ app_dart/lib/src/request_handlers/scheduler/backfill_grid.dart
@@ -11,8 +11,9 @@ import 'dart:convert';
 import 'package:meta/meta.dart';
 
 import '../../model/ci_yaml/target.dart';
+import '../../model/commit_ref.dart';
 import '../../model/firestore/task.dart';
-import '../../service/luci_build_service/commit_task_ref.dart';
+import '../../model/task_ref.dart';
 import '../../service/luci_build_service/pending_task.dart';
 import '../../service/scheduler/policy.dart';
