+++ app_dart/lib/src/request_handlers/scheduler/batch_backfiller.dart
@@ -14,7 +14,6 @@ import '../../model/ci_yaml/ci_yaml.dart';
 import '../../model/firestore/task.dart' as fs;
 import '../../request_handling/exceptions.dart';
 import '../../service/firestore/commit_and_tasks.dart';
-import '../../service/luci_build_service/commit_task_ref.dart';
 import '../../service/luci_build_service/pending_task.dart';
 import '../../service/scheduler/ci_yaml_fetcher.dart';
 import '../../service/scheduler/policy.dart';
@@ -112,10 +111,7 @@ final class BatchBackfiller extends RequestHandler {
 
       grid = BackfillGrid.from([
         for (final CommitAndTasks(:commit, :tasks) in fsGrid)
-          (
-            CommitRef.fromFirestore(commit),
-            [...tasks.map(TaskRef.fromFirestore)],
-          ),
+          (commit.toRef(), [...tasks.map((t) => t.toRef())]),
       ], tipOfTreeTargets: totTargets);
     }
     log.debug('Built a grid of ${grid.eligibleTasks.length} target columns');
