+++ app_dart/lib/src/service/luci_build_service.dart
@@ -18,13 +18,13 @@ import 'package:meta/meta.dart';
 import '../../cocoon_service.dart';
 import '../foundation/github_checks_util.dart';
 import '../model/ci_yaml/target.dart';
+import '../model/commit_ref.dart';
 import '../model/firestore/pr_check_runs.dart' as fs;
 import '../model/firestore/task.dart' as fs;
 import '../model/github/checks.dart' as cocoon_checks;
 import 'exceptions.dart';
 import 'luci_build_service/build_tags.dart';
 import 'luci_build_service/cipd_version.dart';
-import 'luci_build_service/commit_task_ref.dart';
 import 'luci_build_service/engine_artifacts.dart';
 import 'luci_build_service/pending_task.dart';
 import 'luci_build_service/user_data.dart';
