+++ app_dart/test/service/luci_build_service/schedule_merge_group_builds_test.dart
@@ -7,7 +7,6 @@ import 'package:cocoon_server_test/test_logging.dart';
 import 'package:cocoon_service/src/service/cache_service.dart';
 import 'package:cocoon_service/src/service/config.dart';
 import 'package:cocoon_service/src/service/luci_build_service.dart';
-import 'package:cocoon_service/src/service/luci_build_service/commit_task_ref.dart';
 import 'package:cocoon_service/src/service/luci_build_service/user_data.dart';
 import 'package:mockito/mockito.dart';
 import 'package:path/path.dart' as p;
@@ -93,7 +92,7 @@ void main() {
       branch: 'gh-readonly-queue/master/pr-1234-abcd',
     );
     await luci.scheduleMergeGroupBuilds(
-      commit: CommitRef.fromFirestore(commit),
+      commit: commit.toRef(),
       targets: [
         generateTarget(1, slug: Config.flutterSlug, properties: {'os': 'abc'}),
         generateTarget(2, slug: Config.flutterSlug, properties: {'os': 'abc'}),
@@ -146,7 +145,7 @@ void main() {
       branch: 'gh-readonly-queue/master/pr-1234-abcd',
     );
     await luci.scheduleMergeGroupBuilds(
-      commit: CommitRef.fromFirestore(commit),
+      commit: commit.toRef(),
       targets: [
         generateTarget(1, slug: Config.flutterSlug, properties: {'os': 'abc'}),
         generateTarget(2, slug: Config.flutterSlug, properties: {'os': 'abc'}),
